Create .bat files
Added a .bat file which run the SWT Applets to each project.

To run one applet it is enough to download the folder with the JAR and
BAT files and run the .bat file.
It is possible to create a direct access of that BAT file and put it in
any folder of your system.