Added new Project
ROFarm is a simulator of simple farms (following the schemes of KoeRO
farming facility.
It simulates a farm and allow to see the result of apply some food with
some initial children.
The UI is based on SWT, and that is why it is in this repository.

More info for the simulation, please contact the owner of the
repository.